+++ ts/features/wallet/onboarding/cobadge/screens/ActivateBpdOnNewCoBadgeScreen.tsx
